+++ bindings_node/test/Conversations.test.ts
@@ -22,9 +22,9 @@ describe('Conversations', () => {
     const user = createUser()
     const client = await createRegisteredClient(user)
 
-    expect((await client.conversations().list()).length).toBe(0)
-    expect((await client.conversations().listDms()).length).toBe(0)
-    expect((await client.conversations().listGroups()).length).toBe(0)
+    expect(client.conversations().list().length).toBe(0)
+    expect(client.conversations().listDms().length).toBe(0)
+    expect(client.conversations().listGroups().length).toBe(0)
   })
 
   it('should create a group chat', async () => {
@@ -67,22 +67,22 @@ describe('Conversations', () => {
 
     expect(group.consentState()).toBe(ConsentState.Allowed)
 
-    const group1 = await client1.conversations().list()
+    const group1 = client1.conversations().list()
     expect(group1.length).toBe(1)
     expect(group1[0].id).toBe(group.id)
-    expect((await client1.conversations().listDms()).length).toBe(0)
-    expect((await client1.conversations().listGroups()).length).toBe(1)
+    expect(client1.conversations().listDms().length).toBe(0)
+    expect(client1.conversations().listGroups().length).toBe(1)
 
-    expect((await client2.conversations().list()).length).toBe(0)
+    expect(client2.conversations().list().length).toBe(0)
 
     await client2.conversations().sync()
 
-    const group2 = await client2.conversations().list()
+    const group2 = client2.conversations().list()
     expect(group2.length).toBe(1)
     expect(group2[0].id).toBe(group.id)
 
-    expect((await client2.conversations().listDms()).length).toBe(0)
-    expect((await client2.conversations().listGroups()).length).toBe(1)
+    expect(client2.conversations().listDms().length).toBe(0)
+    expect(client2.conversations().listGroups().length).toBe(1)
   })
 
   it('should create a group with custom permissions', async () => {
@@ -213,25 +213,25 @@ describe('Conversations', () => {
 
     expect(group.consentState()).toBe(ConsentState.Allowed)
 
-    const group1 = await client1.conversations().list()
+    const group1 = client1.conversations().list()
     expect(group1.length).toBe(1)
     expect(group1[0].id).toBe(group.id)
     expect(group1[0].dmPeerInboxId()).toBe(client2.inboxId())
 
-    expect((await client1.conversations().listDms()).length).toBe(1)
-    expect((await client1.conversations().listGroups()).length).toBe(0)
+    expect(client1.conversations().listDms().length).toBe(1)
+    expect(client1.conversations().listGroups().length).toBe(0)
 
-    expect((await client2.conversations().list()).length).toBe(0)
+    expect(client2.conversations().list().length).toBe(0)
 
     await client2.conversations().sync()
 
-    const group2 = await client2.conversations().list()
+    const group2 = client2.conversations().list()
     expect(group2.length).toBe(1)
     expect(group2[0].id).toBe(group.id)
     expect(group2[0].dmPeerInboxId()).toBe(client1.inboxId())
 
-    expect((await client2.conversations().listDms()).length).toBe(1)
-    expect((await client2.conversations().listGroups()).length).toBe(0)
+    expect(client2.conversations().listDms().length).toBe(1)
+    expect(client2.conversations().listGroups().length).toBe(0)
 
     const dm1 = client1.conversations().findDmByTargetInboxId(client2.inboxId())
     expect(dm1).toBeDefined()
@@ -487,15 +487,15 @@ describe('Conversations', () => {
 
     const groups2 = client2.conversations()
     await groups2.sync()
-    const groupsList2 = await groups2.list()
+    const groupsList2 = groups2.list()
 
     const groups3 = client3.conversations()
     await groups3.sync()
-    const groupsList3 = await groups3.list()
+    const groupsList3 = groups3.list()
 
-    const groups4 = await client4.conversations()
+    const groups4 = client4.conversations()
     await groups4.sync()
-    const groupsList4 = await groups4.list()
+    const groupsList4 = groups4.list()
 
     const message1 = await groupsList2[0].send(encodeTextMessage('gm!'))
     const message2 = await groupsList3[0].send(encodeTextMessage('gm2!'))
@@ -530,15 +530,15 @@ describe('Conversations', () => {
 
     const groups2 = client2.conversations()
     await groups2.sync()
-    const groupsList2 = await groups2.list()
+    const groupsList2 = groups2.list()
 
     const groups3 = client3.conversations()
     await groups3.sync()
-    const groupsList3 = await groups3.list()
+    const groupsList3 = groups3.list()
 
-    const groups4 = await client4.conversations()
+    const groups4 = client4.conversations()
     await groups4.sync()
-    const groupsList4 = await groups4.list()
+    const groupsList4 = groups4.list()
 
     await groupsList4[0].send(encodeTextMessage('gm3!'))
     const message1 = await groupsList2[0].send(encodeTextMessage('gm!'))
@@ -573,15 +573,15 @@ describe('Conversations', () => {
 
     const groups2 = client2.conversations()
     await groups2.sync()
-    const groupsList2 = await groups2.list()
+    const groupsList2 = groups2.list()
 
     const groups3 = client3.conversations()
     await groups3.sync()
-    const groupsList3 = await groups3.list()
+    const groupsList3 = groups3.list()
 
-    const groups4 = await client4.conversations()
+    const groups4 = client4.conversations()
     await groups4.sync()
-    const groupsList4 = await groups4.list()
+    const groupsList4 = groups4.list()
 
     await groupsList2[0].send(encodeTextMessage('gm!'))
     await groupsList3[0].send(encodeTextMessage('gm2!'))